+++ src/plugin/engine/improved_analytical_engine/improved_analytical_engine.cpp
@@ -25,8 +25,8 @@ double Analytical_engine::check_collision(double future_time) {
     if (double_are_same(future_time, 0)) return -1;
 
     Sim_config config = arena_ptr->get_config();
-    int max_x = config.get_arena_max_x();
-    int max_y = config.get_arena_max_y();
+    double max_x = config.get_arena_max_x();
+    double max_y = config.get_arena_max_y();
     int num_robots = config.get_num_robots();
     // std::cout << "engine start" << std::endl;
     for (int i = 0; i < num_robots; i++) {
@@ -151,8 +151,8 @@ double Analytical_engine::check_collision(double future_time) {
 }
 
 double Analytical_engine::time_to_out_of_bound(position2d_t pos, double v,
-                                               int radius, int x_max,
-                                               int y_max) {
+                                               double radius, double x_max,
+                                               double y_max) {
     double v_x = cos(((double)pos.theta) / 180 * M_PI) * v;
     double v_y = sin(((double)pos.theta) / 180 * M_PI) * v;
     double min_time = DBL_MAX;
@@ -188,7 +188,7 @@ double Analytical_engine::time_to_out_of_bound(position2d_t pos, double v,
 
 double Analytical_engine::time_to_collide(position2d_t pos_1, double v_1,
                                           position2d_t pos_2, double v_2,
-                                          int radius) {
+                                          double radius) {
     double v_1_x = cos(((double)pos_1.theta) / 180 * M_PI) * v_1;
     double v_1_y = sin(((double)pos_1.theta) / 180 * M_PI) * v_1;
     double v_2_x = cos(((double)pos_2.theta) / 180 * M_PI) * v_2;
